Hooked up signals between UI and GameManager
Connected the Bake button that appears above the Furnace to the GameManager.

This causes the GameManager to record a single new Biscuits as having been made.

A similar implementation has been made that allows the End Turn button to tell the GameManager to update the current turn by advancing a single turn

This implementation will have to be modified to allow for more dynamic amounts of Biscuits to be created.

BiscuitLabel a UI element in the top left checks the GameManager and displays the current turn and the current number of biscuits